@@ -13,9 +13,9 @@ dependencies:
   cupertino_icons: ^1.0.6
   provider: ^6.1.1
   go_router: ^12.1.1
-  dio: ^5.4.0
+  dio: ^5.8.0+1
   shared_preferences: ^2.2.2
-  path_provider: ^2.1.1
+  path_provider: ^2.1.5
   just_audio: ^0.9.36
   just_audio_background: ^0.0.1-beta.15
   just_audio_windows: ^0.2.2
@@ -26,8 +26,8 @@ dependencies:
   lottie: ^2.7.0
   path_provider_windows: ^2.3.0
   shared_preferences_windows: ^2.4.1
-  crypto: ^3.0.3
-  webview_flutter: ^4.4.2
+  crypto: ^3.0.6
+  webview_flutter: ^4.9.0
   webview_flutter_android: ^3.13.1
   webview_flutter_wkwebview: ^3.10.0
   audio_service: ^0.18.12
@@ -36,7 +36,7 @@ dependencies:
   get_it: ^7.6.0
   injectable: ^2.1.2
   path: ^1.8.3
-  url_launcher: ^6.1.14
+  url_launcher: ^6.3.1
   intl: ^0.18.1
   window_manager: ^0.3.9
   hive: ^2.2.3
@@ -44,11 +44,12 @@ dependencies:
   flutter_svg: ^2.0.9
   flutter_easyloading: ^3.0.5
   animations: ^2.0.8
-  permission_handler: ^11.0.1
+  permission_handler: ^12.0.0+1
   fftea: ^1.5.0
   qr_flutter: ^4.1.0
   encrypt: ^5.0.3
   pointycastle: ^3.9.1
+  http: ^1.3.0
 
 dev_dependencies:
   flutter_test:
